+++ views/bootstrap/thead.blade.php
@@ -1,15 +1,13 @@
 <thead class="table-header">
     {{-- rows number / search --}}
-    @if(! $table->searchableColumns->isEmpty()
-        || $table->rowsNumberSelectionActivation)
+    @if(count($table->sortableColumns) || count($table->searchableColumns) || $table->rowsNumberSelectionActivation)
         <tr {{ classTag($table->trClasses) }}>
-            <td {{ classTag('border-0', 'p-0', $table->tdClasses) }}
-                colspan="{{ $table->columnsCount() + ($table->isRouteDefined('edit') 
-                    || $table->isRouteDefined('destroy') ? 1 : 0) }}">
-                <div class="d-flex justify-content-between flex-wrap">
-                    {{-- rows number selector --}}
-                    @if($table->rowsNumberSelectionActivation)
-                        <div class="flex-fill p-3 rows-number-selector">
+            <td {{ classTag('border-0', 'py-4', $table->tdClasses) }}
+                colspan="{{ $table->columnsCount() + ($table->isRouteDefined('edit') || $table->isRouteDefined('destroy') ? 1 : 0) }}">
+                <div class="row">
+                    {{-- rows number selection --}}
+                    <div class="col-sm-12 col-lg-4 rows-number-selection">
+                        @if($table->rowsNumberSelectionActivation)
                             <form role="form" method="GET" action="{{ $table->route('index') }}">
                                 <input type="hidden" name="search" value="{{ $table->request->search }}">
                                 <input type="hidden" name="sortBy" value="{{ $table->request->sortBy }}">
@@ -38,11 +36,13 @@
                                     </div>
                                 </div>
                             </form>
-                        </div>
-                    @endif
-                    {{-- search --}}
-                    @if(! $table->searchableColumns->isEmpty())
-                        <div class="flex-fill p-3 search-bar">
+                        @endif
+                    </div>
+                    {{-- spacer --}}
+                    <div class="spacer col-sm-2"></div>
+                    {{-- searching --}}
+                    <div class="col-sm-12 col-lg-6 searching">
+                        @if(count($table->searchableColumns))
                             <form role="form" method="GET" action="{{ $table->route('index') }}">
                                 <input type="hidden" name="rows" value="{{ $table->request->rows }}">
                                 <input type="hidden" name="sortBy" value="{{ $table->request->sortBy }}">
@@ -64,8 +64,13 @@
                                            aria-label="@lang('laravel-table::laravel-table.search') {{ $table->searchableTitles() }}">
                                     @if($table->request->search)
                                         <div class="input-group-append">
-                                            <a class="input-group-text btn btn-link text-danger"
-                                               href="{{ $table->route('index', ['search' => null, 'rows' => $table->request->rows, 'sortBy' => $table->request->sortBy, 'sortDir' => $table->request->sortDir]) }}"
+                                            <a class="input-group-text btn btn-link text-danger cancel-search"
+                                               href="{{ $table->route('index', array_merge([
+                                                    'search'    => null,
+                                                    'rows'      => $table->request->rows,
+                                                    'sortBy'    => $table->request->sortBy,
+                                                    'sortDir'   => $table->request->sortDir
+                                                ], $table->appendedValues)) }}"
                                                title="@lang('laravel-table::laravel-table.cancelSearch')">
                                                 <span>{!! config('laravel-table.icon.cancel') !!}</span>
                                             </a>
@@ -81,8 +86,8 @@
                                     @endif
                                 </div>
                             </form>
-                        </div>
-                    @endif
+                        @endif
+                    </div>
                 </div>
             </td>
         </tr>
@@ -93,20 +98,18 @@
             <th {{ classTag('border-0', $table->thClasses) }} scope="col">
                 @if($column->isSortable)
                     <a href="{{ $table->route('index', array_merge([
-                                'sortBy' => $column->databaseDefaultColumn,
-                                'sortDir' => $table->request->sortDir === 'desc' ? 'asc' : 'desc',
-                                'search'   => $table->request->search,
-                                'rows'    => $table->request->rows
-                            ], $table->appendedValues)) }}"
+                            'sortBy'    => $column->databaseDefaultColumn,
+                            'sortDir'   => $table->request->sortDir === 'desc' ? 'asc' : 'desc',
+                            'search'    => $table->request->search,
+                            'rows'      => $table->request->rows
+                        ], $table->appendedValues)) }}"
                        title="{{ $column->title }}">
-                        @if($table->request->sortBy === $column->databaseDefaultColumn 
-                            && $table->request->sortDir === 'asc')
-                            <span class="sort">
+                        @if($table->request->sortBy === $column->databaseDefaultColumn && $table->request->sortDir === 'asc')
+                            <span class="sort asc">
                                 {!! config('laravel-table.icon.sortAsc') !!}
                             </span>
-                        @elseif($table->request->sortBy === $column->databaseDefaultColumn 
-                            && $table->request->sortDir === 'desc')
-                            <span class="sort">
+                        @elseif($table->request->sortBy === $column->databaseDefaultColumn && $table->request->sortDir === 'desc')
+                            <span class="sort desc">
                                 {!! config('laravel-table.icon.sortDesc') !!}
                             </span>
                         @else