+++ CHANGELOG.md
@@ -2,9 +2,7 @@
 
 ## [1.0.4](https://github.com/Okipa/laravel-table/releases/tag/1.0.4)
 2019-02-17
-- `->appends()` does now also merge appended keys values to request attributes, in order to use the appended to the pagination, but also to the rows number selection, searching and sorting actions.
-- Replaced `col-[breakpoint]` thead templating by `flex`. 
-- Replaced `col-[breakpoint]` tfoot templating by `flex`. 
+- `->appends()` does now also add appended attributes to search canceling and sorting actions.
 
 ## [1.0.3](https://github.com/Okipa/laravel-table/releases/tag/1.0.3)
 2019-02-15